+++ wrapper/wrapper_writer.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "..\common\rawaccel.hpp"
+
+struct writer {
+	void writeToDriver(rawaccel::mouse_modifier* modifier);
+};
